+++ otx/__init__.py
@@ -5,3 +5,28 @@
 
 __version__ = "1.0.0"
 # NOTE: Sync w/ otx/api/usecases/exportable_code/demo/requirements.txt
+
+MMCLS_AVAILABLE = True
+MMDET_AVAILABLE = True
+MMSEG_AVAILABLE = True
+MMACTION_AVAILABLE = True
+
+try:
+    import mmcls
+except ImportError:
+    MMCLS_AVAILABLE = False
+
+try:
+    import mmdet
+except ImportError:
+    MMDET_AVAILABLE = False
+
+try:
+    import mmseg
+except ImportError:
+    MMSEG_AVAILABLE = False
+
+try:
+    import mmaction
+except ImportError:
+    MMACTION_AVAILABLE = False
