+++ otx/mpa/det/inferrer.py
@@ -20,9 +20,9 @@ from otx.algorithms.common.adapters.mmcv.utils import (
 from otx.algorithms.detection.adapters.mmdet.data import ImageTilingDataset
 from otx.mpa.modules.hooks.recording_forward_hooks import (
     ActivationMapHook,
-    DetSaliencyMapHook,
     FeatureVectorHook,
 )
+from otx.mpa.modules.hooks.det_saliency_map_hook import DetSaliencyMapHook
 from otx.mpa.registry import STAGES
 from otx.mpa.utils.logger import get_logger
