+++ otx/mpa/modules/models/detectors/custom_yolox_detector.py
@@ -114,10 +114,8 @@ if is_mmdeploy_enabled():
     from mmdeploy.core import FUNCTION_REWRITER, mark
     from mmdeploy.utils import is_dynamic_shape
 
-    from otx.mpa.modules.hooks.recording_forward_hooks import (
-        DetSaliencyMapHook,
-        FeatureVectorHook,
-    )
+    from otx.mpa.modules.hooks.recording_forward_hooks import FeatureVectorHook
+    from otx.mpa.modules.hooks.det_saliency_map_hook import DetSaliencyMapHook
 
     @FUNCTION_REWRITER.register_rewriter(
         "otx.mpa.modules.models.detectors.custom_yolox_detector.CustomYOLOX.simple_test"
