+++ requirements/segmentation.txt
@@ -4,3 +4,4 @@ mmcv-full==1.7.0
 mmsegmentation==0.30.0
 scikit-image
 otxdeploy==0.12.1  # FIXME: mmdeploy==? when PyPI packages are available
+timm
